cleaned up the utilities folder by splitting up code into seperate files
